fix: :bug: Fixed connections to the 'keycloak' endpoints fail with "502 Bad Gateway".
keyclock is listening on port 8090 instead on port 8080.The proxy sends request to 8080 instead of 8090 after the change .  The keyclock serivce in docker-compose file  mapping is done  between port 8080 on the host(keyclock) and port 8080 on the container.

The issue "Connections to the 'keycloak' endpoints fail with "502 Bad Gateway" is fixed.